+++ lib/onboardable/list/navigation.rb
@@ -8,30 +8,48 @@ module Onboardable
       #
       # @return [Step, nil] The next step in the list or nil if the current step is the last one.
       def next_step
-        current_index = step_index!(current_step)
+        current_index = step_index(current_step)
         steps[current_index.next]
+      end
+
+      # Checks if the specified step is the next step in the onboarding process.
+      #
+      # @param step [Step] The step to check.
+      # @return [Boolean] True if the specified step is the next step, false otherwise.
+      def next_step?(step)
+        next_step == step
       end
 
       # Moves the current step pointer to the next step in the onboarding process.
       #
-      # @raise [LastStepError] if the current step is the last step and there is no next step to move to.
+      # @return [Step] The next step in the list.
+      # @raise [LastStepError] If the current step is the last step and there is no next step to move to.
       def next_step!
-        self.current_step = next_step || raise(LastStepError.new(current_step, steps.map(&:to_str)))
+        self.current_step = next_step || last_step_error!
       end
 
       # Returns the previous step in the onboarding process.
       #
       # @return [Step, nil] The previous step in the list or nil if the current step is the first one.
       def prev_step
-        current_index = step_index!(current_step)
+        current_index = step_index(current_step)
         current_index.positive? ? steps[current_index.pred] : nil
+      end
+
+      # Checks if the specified step is the previous step in the onboarding process.
+      #
+      # @param step [Step] The step to check.
+      # @return [Boolean] True if the specified step is the previous step, false otherwise.
+      def prev_step?(step)
+        prev_step == step
       end
 
       # Moves the current step pointer to the previous step in the onboarding process.
       #
-      # @raise [FirstStepError] if the current step is the first step and there is no previous step to move to.
+      # @return [Step] The previous step in the list.
+      # @raise [FirstStepError] If the current step is the first step and there is no previous step to move to.
       def prev_step!
-        self.current_step = prev_step || raise(FirstStepError.new(current_step, steps.map(&:to_str)))
+        self.current_step = prev_step || first_step_error!
       end
 
       # Checks if the specified step is the first step in the onboarding process.
@@ -39,7 +57,7 @@ module Onboardable
       # @param step [Step] The step to check (defaults to the current step if not specified).
       # @return [Boolean] True if the specified step is the first step, false otherwise.
       def first_step?(step = current_step)
-        step == first_step
+        first_step == step
       end
 
       # Checks if the specified step is the last step in the onboarding process.
@@ -47,7 +65,7 @@ module Onboardable
       # @param step [Step] The step to check (defaults to the current step if not specified).
       # @return [Boolean] True if the specified step is the last step, false otherwise.
       def last_step?(step = current_step)
-        step == last_step
+        last_step == step
       end
 
       # Retrieves the first step in the onboarding process.
@@ -57,11 +75,35 @@ module Onboardable
         steps.fetch(0)
       end
 
+      # Checks if the specified step is the current step in the onboarding process.
+      #
+      # @param step [Step] The step to check.
+      # @return [Boolean] True if the specified step is the current step, false otherwise.
+      def current_step?(step)
+        current_step == step
+      end
+
       # Retrieves the last step in the onboarding process.
       #
       # @return [Step] The last step in the list.
       def last_step
         steps.fetch(-1)
+      end
+
+      private
+
+      # Raises a FirstStepError indicating the current step is the first step in the onboarding process.
+      #
+      # @raise [FirstStepError] The error indicating the current step is the first step.
+      def first_step_error!
+        raise FirstStepError.new(current_step, steps.map(&:to_str))
+      end
+
+      # Raises a LastStepError indicating the current step is the last step in the onboarding process.
+      #
+      # @raise [LastStepError] The error indicating the current step is the last step.
+      def last_step_error!
+        raise LastStepError.new(current_step, steps.map(&:to_str))
       end
     end
   end